% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{autor}
\alias{autor}
\title{Dataset from David Autor (2003).}
\format{
A data frame with 1200 rows and 598 variables:
}
\source{
\url{https://economics.mit.edu/faculty/dautor/data/autor03}
}
\usage{
autor
}
\description{
Data from:
Outsourcing at Will: The Contribution of Unjust Dismissal Doctrine to the Growth of Employment Outsourcing
David Autor
January 2003, Journal of Labor Economics, 2003, 21(1)
}
\keyword{datasets}
