% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_impute.R
\name{did_impute}
\alias{did_impute}
\title{\strong{Work in Progress} Treatment effect estimation and pre-trend testing in staggered adoption diff-in-diff designs with an imputation approach of Borusyak, Jaravel, and Spiess (2021)}
\usage{
did_impute(
  data,
  yname,
  gname,
  tname,
  idname,
  first_stage_formula,
  treat_var,
  weights
)
}
\description{
\strong{Work in Progress} Treatment effect estimation and pre-trend testing in staggered adoption diff-in-diff designs with an imputation approach of Borusyak, Jaravel, and Spiess (2021)
}
