% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honest_did.R
\name{get_honestdid_obj_did2s}
\alias{get_honestdid_obj_did2s}
\title{get_honestdid_obj_did2s}
\usage{
get_honestdid_obj_did2s(est, coef_name = "rel_year")
}
\arguments{
\item{est}{A \code{fixest} object, likely from \code{did2s}.}

\item{coef_name}{Character. The name of the event-study relative-year
variable name, from \code{i(rel_year)}.}
}
\value{
A list containing the vector of event-study coefficients \code{beta}, the
variance-covariance matrix of \code{beta}, \code{V}, and a vector of relative years,
\code{event_time}.
}
\description{
a helper function that takes a fixest feols object (likely
from \code{did2s}) that plugs into \code{honest_did}. Note this function assumes
the event study coefficients are using \code{i()} syntax, e.g. \code{i(rel_year)}.
This should also work for a TWFE event-study model estimated by \code{feols}.
}
