% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_study.R
\name{plot_event_study}
\alias{plot_event_study}
\title{Plot results of \code{\link[=event_study]{event_study()}}}
\usage{
plot_event_study(out, seperate = TRUE, horizon = NULL)
}
\arguments{
\item{out}{Tibble from \code{\link[=event_study]{event_study()}}}

\item{seperate}{Logical. Should the estimators be on seperate plots? Default is TRUE.}

\item{horizon}{Numeric. Vector of length 2. First element is min and second element is max of event_time to plot}
}
\value{
ggplot object that can be fully customized
}
\description{
Plot results of \code{\link[=event_study]{event_study()}}
}
\examples{
data(df_het, package = "did2s")
out = event_study(
  data = df_het, yname = "dep_var", idname = "unit",
    tname = "year", gname = "g"
)
plot_event_study(out)

}
