% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sparse_model_matrix}
\alias{sparse_model_matrix}
\title{Make a sparse_model_matrix for fixest estimate}
\usage{
sparse_model_matrix(data, fixest)
}
\description{
Make a sparse_model_matrix for fixest estimate
}
