% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_kyle.R
\name{theme_kyle}
\alias{theme_kyle}
\title{Kyle Butts' custom ggplot theme}
\usage{
theme_kyle(
  title_pos = "center",
  axis_title_pos = "left",
  slides = FALSE,
  has_subtitle = FALSE,
  base_size = 14,
  ...
)
}
\arguments{
\item{title_pos}{String. Either "center" or "left" to align title and subtitle center of left respectively}

\item{axis_title_pos}{String. Either "center" or "left" to align title and subtitle center or left/top respectively}

\item{slides}{Logical. if TRUE, use background #ECECEC or not. Default is FALSE.}

\item{has_subtitle}{Logical. Adjusts margins when subtitle is used. Default is FALSE.}

\item{base_size}{Numeric. Base size of font. Everything else is relative to this font.}
}
\description{
Kyle Butts' custom ggplot theme
}
