% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did2s.R
\name{did2s}
\alias{did2s}
\title{Calculate two-stage difference-in-differences following Gardner (2021)}
\usage{
did2s(
  data,
  yname,
  first_stage_formula,
  treat_formula,
  treat_var,
  cluster_var,
  weights = NULL,
  bootstrap = FALSE,
  n_bootstraps = 250
)
}
\arguments{
\item{data}{The dataframe containing all the variables}

\item{yname}{Outcome variable}

\item{first_stage_formula}{Fixed effects and other covariates you want to residualize with in first stage, use i() for fixed effects., following fixest::feols.}

\item{treat_formula}{Second stage, these should be the treatment indicator(s) (e.g. treatment variable or es leads/lags), use i() for factor variables, following fixest::feols.}

\item{treat_var}{A variable that = 1 if treated, = 0 otherwise}

\item{cluster_var}{What variable to cluster standard errors. This can be IDs or a higher aggregate level (state for example)}

\item{weights}{Optional variable to run a weighted first- and second-stage regressions}

\item{bootstrap}{Should standard errors be calculated using bootstrap? Default is FALSE.}

\item{n_bootstraps}{How many bootstraps to run. Default is 250}
}
\value{
fixest::feols point estimate with adjusted standard errors (either by formula or by bootstrap)
}
\description{
Calculate two-stage difference-in-differences following Gardner (2021)
}
\examples{
# Load Example Dataset
data("df_hom")

# Static
static <- did2s(df_hom, yname = "dep_var", first_stage_formula = "i(state) + i(year)", treat_formula = "i(treat)", treat_var = "treat", cluster_var = "state")
fixest::esttable(static)

# Event-Study
es <- did2s(df_hom, yname = "dep_var", first_stage_formula = "i(state) + i(year)", treat_formula = "i(rel_year)", treat_var = "treat", cluster_var = "state")
fixest::esttable(es)

# Castle Data
castle <- haven::read_dta("https://github.com/scunning1975/mixtape/raw/master/castle.dta")

did2s(
	data = castle,
	yname = "l_homicide",
	first_stage_formula = ~ i(sid) + i(year),
	treat_formula = ~ i(post, ref=0),
	treat_var = "post",
	cluster_var = "state", weights = "popwt"
)

}
