% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did2s.R
\name{did2s}
\alias{did2s}
\title{Calculate two-stage difference-in-differences following Gardner (2021)}
\usage{
did2s(
  data,
  yname,
  first_stage,
  second_stage,
  treatment,
  cluster_var,
  weights = NULL,
  bootstrap = FALSE,
  n_bootstraps = 250,
  verbose = TRUE
)
}
\arguments{
\item{data}{The dataframe containing all the variables}

\item{yname}{Outcome variable}

\item{first_stage}{Fixed effects and other covariates you want to residualize
with in first stage.
Formula following \code{\link[fixest:feols]{fixest::feols}}.
Fixed effects specified after "\code{|}".}

\item{second_stage}{Second stage, these should be the treatment indicator(s)
(e.g. treatment variable or event-study leads/lags).
Formula following \code{\link[fixest:feols]{fixest::feols}}.
Use \code{i()} for factor variables, see \code{\link[fixest:i]{fixest::i}}.}

\item{treatment}{A variable that = 1 if treated, = 0 otherwise}

\item{cluster_var}{What variable to cluster standard errors. This can be IDs
or a higher aggregate level (state for example)}

\item{weights}{Optional. Variable name for regression weights.}

\item{bootstrap}{Optional. Should standard errors be calculated using bootstrap?
Default is \code{FALSE}.}

\item{n_bootstraps}{Optional. How many bootstraps to run.
Default is \code{250}.}

\item{verbose}{Optional. Logical. Should information about the two-stage
procedure be printed back to the user?
Default is \code{TRUE}.}
}
\value{
fixest::feols point estimate with adjusted standard errors (either by formula or by bootstrap)
}
\description{
Calculate two-stage difference-in-differences following Gardner (2021)
}
\section{Examples}{


Load example dataset which has two treatment groups and homogeneous treatment effects\if{html}{\out{<div class="r">}}\preformatted{# Load Example Dataset
data("df_hom")
}\if{html}{\out{</div>}}

You can run a static TWFE fixed effect model for a simple treatment indicator\if{html}{\out{<div class="r">}}\preformatted{static <- did2s(df_hom,
    yname = "dep_var", treatment = "treat", cluster_var = "state",
    first_stage = "i(state) + i(year)",
    second_stage = "i(treat, ref=FALSE)")
#> 
#> ── Two-stage Difference-in-Differences ─────────────────────────────────────────
#> → Running with first stage formula `~ i(state) + i(year)` and second stage formula `~ i(treat, ref=FALSE)`
#> → The indicator variable that denotes when treatment is on is `treat`
#> → Standard errors will be clustered by `state`

fixest::esttable(static)
#>                            static
#> Dependent Var.:           dep_var
#>                                  
#> treat = TRUE    1.999*** (0.0627)
#> _______________ _________________
#> S.E. type                  Custom
#> Observations               31,000
#> R2                        0.23037
#> Adj. R2                   0.23037
}\if{html}{\out{</div>}}

Or you can use relative-treatment indicators to estimate an event study estimate\if{html}{\out{<div class="r">}}\preformatted{es <- did2s(df_hom,
    yname = "dep_var", treatment = "treat", cluster_var = "state",
    first_stage = "i(state) + i(year)",
    second_stage = "i(rel_year, ref=c(-1, Inf))")
#> 
#> ── Two-stage Difference-in-Differences ─────────────────────────────────────────
#> → Running with first stage formula `~ i(state) + i(year)` and second stage formula `~ i(rel_year, ref=c(-1, Inf))`
#> → The indicator variable that denotes when treatment is on is `treat`
#> → Standard errors will be clustered by `state`

fixest::esttable(es)
#>                                es
#> Dependent Var.:           dep_var
#>                                  
#> rel_year = -20   -0.0636 (0.0760)
#> rel_year = -19    0.0281 (0.0772)
#> rel_year = -18    0.0048 (0.0771)
#> rel_year = -17   -0.0064 (0.0857)
#> rel_year = -16   -0.0185 (0.0692)
#> rel_year = -15   -0.0448 (0.0974)
#> rel_year = -14    0.0849 (0.0828)
#> rel_year = -13   -0.0408 (0.0897)
#> rel_year = -12   -0.1028 (0.0930)
#> rel_year = -11    0.0365 (0.0660)
#> rel_year = -10    0.0266 (0.0520)
#> rel_year = -9    -0.0238 (0.0469)
#> rel_year = -8    -0.0025 (0.0522)
#> rel_year = -7    -0.0131 (0.0508)
#> rel_year = -6   -0.0959. (0.0544)
#> rel_year = -5     0.0059 (0.0523)
#> rel_year = -4    -0.0794 (0.0604)
#> rel_year = -3    -0.0274 (0.0485)
#> rel_year = -2     0.0093 (0.0535)
#> rel_year = 0    2.092*** (0.0708)
#> rel_year = 1    1.831*** (0.0841)
#> rel_year = 2    1.961*** (0.0690)
#> rel_year = 3    1.979*** (0.0798)
#> rel_year = 4    1.925*** (0.0896)
#> rel_year = 5    2.013*** (0.0973)
#> rel_year = 6    2.006*** (0.0980)
#> rel_year = 7    2.000*** (0.0890)
#> rel_year = 8    1.950*** (0.0909)
#> rel_year = 9    2.096*** (0.0710)
#> rel_year = 10   2.048*** (0.1173)
#> rel_year = 11   1.919*** (0.1071)
#> rel_year = 12   1.916*** (0.1300)
#> rel_year = 13   1.940*** (0.1309)
#> rel_year = 14   1.999*** (0.1435)
#> rel_year = 15   2.211*** (0.1210)
#> rel_year = 16   2.154*** (0.1584)
#> rel_year = 17   1.911*** (0.1133)
#> rel_year = 18   2.111*** (0.1338)
#> rel_year = 19   2.088*** (0.1345)
#> rel_year = 20   1.901*** (0.1313)
#> _______________ _________________
#> S.E. type                  Custom
#> Observations               31,000
#> R2                        0.23120
#> Adj. R2                   0.23024
}\if{html}{\out{</div>}}

Here's an example using data from Castle (2013)\if{html}{\out{<div class="r">}}\preformatted{# Castle Data
castle <- haven::read_dta("https://github.com/scunning1975/mixtape/raw/master/castle.dta")

did2s(
	data = castle,
	yname = "l_homicide",
	first_stage = ~ i(sid) + i(year),
	second_stage = ~ i(post, ref=0),
	treatment = "post",
	cluster_var = "state", weights = "popwt"
)
#> 
#> ── Two-stage Difference-in-Differences ─────────────────────────────────────────
#> → Running with first stage formula `~ i(sid) + i(year)` and second stage formula `~ i(post, ref = 0)`
#> → The indicator variable that denotes when treatment is on is `post`
#> → Standard errors will be clustered by `state`
#> OLS estimation, Dep. Var.: l_homicide
#> Observations: 550 
#> Standard-errors: Custom 
#>         Estimate Std. Error t value Pr(>|t|))    
#> post::1 0.075142    0.03538  2.1239  0.034127 *  
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 263.4   Adj. R2: 0.052465
}\if{html}{\out{</div>}}
}

