% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_imputation.R
\name{did_imputation}
\alias{did_imputation}
\title{Treatment effect estimation and pre-trend testing in staggered adoption diff-in-diff designs with an imputation approach of Borusyak, Jaravel, and Spiess (2021)}
\usage{
did_imputation(
  data,
  yname,
  gname,
  tname,
  idname,
  first_stage = NULL,
  weights = NULL,
  wtr = NULL,
  horizon = NULL,
  pretrends = NULL
)
}
\arguments{
\item{data}{A data frame}

\item{yname}{String. Variable name for outcome.}

\item{gname}{String. Variable name for unit-specific date of treatment
(never-treated should be zero or \code{NA})}

\item{tname}{String. Variable name for calendar period}

\item{idname}{String. Variable name for unique unit id}

\item{first_stage}{Formula for Y(0).
Formula following \code{\link[fixest:feols]{fixest::feols}}.
Fixed effects specified after "\code{|}".
If not specified, then just unit and time fixed effects will be used.}

\item{weights}{Estimation weights for observations. This is used in estimating
Y(0) and also augments treatment effect weights.}

\item{wtr}{Character vector of treatment weight names
(see horizon for standard static and event-study weights)}

\item{horizon}{Integer vector of event_time or \code{TRUE}. This only applies if \code{wtr} is left
as \code{NULL}. if specified, weighted averages/sums of treatment effects will be
reported for each of these horizons separately (i.e. tau0 for the treatment
period, tau1 for one period after treatment, etc.).
If \code{TRUE}, all horizons are used.
If \code{wtr} and \code{horizon} are null, then the static treatment effect is calculated.}

\item{pretrends}{Integer vector or \code{TRUE}. Which pretrends to estimate.
If \code{TRUE}, all \code{pretrends} are used.}
}
\description{
Treatment effect estimation and pre-trend testing in staggered adoption diff-in-diff designs with an imputation approach of Borusyak, Jaravel, and Spiess (2021)
}
