% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_imputation.R
\name{did_imputation}
\alias{did_imputation}
\title{\strong{Work in Progress} Treatment effect estimation and pre-trend testing in staggered adoption diff-in-diff designs with an imputation approach of Borusyak, Jaravel, and Spiess (2021)}
\usage{
did_imputation(
  data,
  yname,
  gname,
  tname,
  idname,
  first_stage,
  weights = NULL,
  wtr = NULL,
  horizon = NULL
)
}
\arguments{
\item{data}{A data frame}

\item{yname}{outcome variable}

\item{gname}{variable for unit-specific date of treatment (never-treated should be zero or NA)}

\item{tname}{variable for calendar period}

\item{idname}{variable for unique unit id}

\item{first_stage}{formula for Y(0)}

\item{weights}{estimation weights for observations. This is used in estimating Y(0) and also augments treatment effect weights}

\item{wtr}{character vector of treatment weight names (see horizon/allhorizon for standard event study weights)}

\item{horizon}{vector of event_time. This only applies if wtr is left as NULL. If both wtr and horizon are null, then all values of event_time will be used.}
}
\description{
\strong{Work in Progress} Treatment effect estimation and pre-trend testing in staggered adoption diff-in-diff designs with an imputation approach of Borusyak, Jaravel, and Spiess (2021)
}
