% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_study.R
\name{event_study}
\alias{event_study}
\title{Estimate event-study coefficients using TWFE and 5 proposed improvements.}
\usage{
event_study(
  data,
  yname,
  idname,
  gname,
  tname,
  xformla = NULL,
  horizon = NULL,
  weights = NULL
)
}
\arguments{
\item{data}{The dataframe containing all the variables}

\item{yname}{Variable name for outcome variable}

\item{idname}{Variable name for unique unit id}

\item{gname}{Variable name for unit-specific date of initial treatment (never-treated should be zero or NA)}

\item{tname}{Variable name for calendar period}

\item{xformla}{A formula for the covariates to include in the model. It should be of the form \code{~ X1 + X2}. Default is NULL.}

\item{horizon}{Integer of length two. The first integer is the earliest pre-effect to include and the second is the latest post-effect to include. Default is all horizons.}

\item{weights}{Variable name for estimation weights. This is used in estimating Y(0) and also augments treatment effect weights}
}
\value{
tibble of point estimates for each estimator
}
\description{
Uses the estimation procedures recommended from Borusyak, Jaravell, Spiess (2021); Callaway and Sant'Anna (2020); Gardner (2021); Roth and Sant'Anna (2021); Sun and Abraham (2020)
}
\section{Examples}{

}

